services/api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
python-dotenv==1.0.1
psycopg[binary]==3.1.18
psycopg-pool==3.2.2
redis==5.0.4
SQLAlchemy==2.0.27
alembic==1.13.1
APScheduler==3.10.4
tzdata==2024.1
boto3==1.34.162
sentry-sdk[fastapi]==2.22.0
